LayoutTests/css3/css-supports-document-parser-context-expected.txt

Makes sure that CSS.supports() uses no-quirks mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.test1 is false
PASS document.test2 is false
PASS successfullyParsed is true

TEST COMPLETE
